+++ Kiwi.Json/Serialization/IJsonWriter.cs
@@ -9,7 +9,7 @@ namespace Kiwi.Json.Serialization
         void WriteObject(IDictionary<string, IJsonValue> obj);
         void WriteArray(IList<IJsonValue> array);
         void WriteString(string value);
-        void WriteInteger(int value);
+        void WriteInteger(long value);
         void WriteDouble(double value);
         void WriteDate(DateTime value);
         void WriteBool(bool value);
